♻️  `KBarPortal` keep `KBarProvider` separate
