+++ examples/design-system/src/components/KBar/index.ts
@@ -1,9 +1,10 @@
+export { KBar } from './KBar'
 export { KBarActions } from './KBarActions'
+export { KBarPortal } from './KBarPortal'
 export { KBarSearchResult } from './KBarSearchResult'
 export { KBarSearchResults } from './KBarSearchResults'
 export { KBarSubscriptions } from './KBarSubscriptions'
 export { KBarWindowControls } from './KBarWindowControls'
-export { KBarProviderApp } from './KBarProviderApp'
 
 export {
   KBarPositionerStyle,
